// src/soumu/blocks/datatable/datatable.type.tsx
export interface SortableQueries {
  dataKey: string,
  type: 'asc' | 'desc';
}

/**
 * Sample data table query structure
 * ```js
 * {
 *   page: 1,
 *   size: 20,
 *   sort: {
 *     key: 'data_key',
 *     type: 'asc',
 *   },
 *   filter: {
 *     filter_key_1: 'any type',
 *     filter_key_2: 'any type',
 *     ...
 *   }
 * }
 * ```
 */
export interface DataTableQueries<F = any> {
  /**
   * Default: 1
   */
  page: number;

  /**
   * Default: from API setting.
   */
  size: number;

  /**
   * Default: null
   */
  sort?: SortableQueries;

  /**
   * Default: null
   */
  filter?: F;
}
// ...
export interface LocalSearchHelper {
  paging: <D>(queries: DataTableQueries, allDataItems: D[], isPaging?: { isPaging: boolean }) => {
    dataItems: D[];
    currentPage?: number;
    // TODO: rename
    isPaging?: { isPaging: boolean };
  };
}